Related work items: #25

+++ HaloAPI.build.ps1
@@ -1,6 +1,6 @@
 <#
     .SYNOPSIS
-        Build script for HaloAPI module - uses 'Invoke-Build' https://github.com/nightroman/Invoke-Build
+        Homotechsual portable module build script.
 #>
 Param (
     [String]$Configuration = 'Development',
@@ -32,8 +32,12 @@ if ($CopyModuleFiles) {
     if (-not (Test-Path "$($PSScriptRoot)\Output\$ModuleName")) {
         New-Item -Path "$($PSScriptRoot)\Output\$ModuleName" -ItemType Directory | Out-Null
     }
-    Copy-Item -Path "$($PSScriptRoot)\Classes\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
-    Copy-Item -Path "$($PSScriptRoot)\Data\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
+    if (Test-Path -Path "$($PSScriptRoot)\Classes\") {
+        Copy-Item -Path "$($PSScriptRoot)\Classes\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
+    }
+    if (Test-Path -Path "$($PSScriptRoot)\Data\") {
+        Copy-Item -Path "$($PSScriptRoot)\Data\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
+    }
     Copy-Item -Path "$($PSScriptRoot)\Private\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
     Copy-Item -Path "$($PSScriptRoot)\Public\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
 
@@ -42,8 +46,8 @@ if ($CopyModuleFiles) {
         "$($PSScriptRoot)\LICENSE.md"
         "$($PSScriptRoot)\CHANGELOG.md"
         "$($PSScriptRoot)\README.md"
-        "$($PSScriptRoot)\HaloAPI.psd1"
-        "$($PSScriptRoot)\HaloAPI.psm1"
+        "$($PSScriptRoot)\$ModuleName.psd1"
+        "$($PSScriptRoot)\$ModuleName.psm1"
     ) -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
 }
